character_agent/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.1
pydantic
google-cloud-storage==2.10.0
google-cloud-run==0.10.3
elevenlabs==0.2.24
aiofiles==23.2.1